refactor: 딜러와 플레이어의 공통 로직 추상 클래스로 분리
