+++ src/main/java/domain/participant/Participant.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import domain.card.Card;
 import domain.card.Cards;
@@ -6,14 +6,13 @@ import domain.card.Cards;
 import java.util.Collections;
 import java.util.List;
 
-public class Dealer {
+abstract class Participant {
 
-    private static final int STAY_LOWER_BOUND = 16;
     private static final int BLACKJACK = 21;
 
     private final Cards cards;
 
-    public Dealer() {
+    Participant() {
         this.cards = new Cards();
     }
 
@@ -22,11 +21,7 @@ public class Dealer {
     }
 
     public int calculateScore() {
-        return cards.calculateScore();
-    }
-
-    public boolean isStay() {
-        return calculateScore() > STAY_LOWER_BOUND;
+        return cards.getScore();
     }
 
     public boolean isBust() {
